legacy/python/局域网文件传输服务器_backend/requirements.txt
fastapi
uvicorn[standard]
PyYAML
python-multipart